Update market_log_to_csv.py
I've introduced the underscore (_) prefix to make certain methods and properties private, as they are not meant to be accessed outside the class.
I've removed the commented-out animation code, as it's not operational and will just clutter up the code. If you plan on using it later, consider making it a private method.
self._log_files is loaded at the initialization, meaning if new log files appear in the directory after the initialization, they won't be processed. If the files may appear after the initialization, consider refreshing the log file list inside the _process_log_files method.